Big Integer Issue in StockOrderActions

+++ ecom-model/src/main/java/com/dowhile/StockOrderDetailCustom.java
@@ -13,9 +13,9 @@ public class StockOrderDetailCustom {
 	private String product_name;
 	private String VariantAttributeName;
 	private BigInteger product_current_inventory;
-	private Integer product_destination_current_inventory;
+	private BigInteger product_destination_current_inventory;
 	private BigInteger product_variant_current_inventory;
-	private Integer product_variant_destination_inventory;
+	private BigInteger product_variant_destination_inventory;
 	private boolean IS_PRODUCT;
 	private Integer ORDER_PROD_QTY;
 	private BigDecimal ORDR_SUPPLY_PRICE;
@@ -70,10 +70,10 @@ public class StockOrderDetailCustom {
 	public void setProduct_current_inventory(BigInteger product_current_inventory) {
 		this.product_current_inventory = product_current_inventory;
 	}
-	public Integer getProduct_destination_current_inventory() {
+	public BigInteger getProduct_destination_current_inventory() {
 		return product_destination_current_inventory;
 	}
-	public void setProduct_destination_current_inventory(Integer product_destination_current_inventory) {
+	public void setProduct_destination_current_inventory(BigInteger product_destination_current_inventory) {
 		this.product_destination_current_inventory = product_destination_current_inventory;
 	}
 	public BigInteger getProduct_variant_current_inventory() {
@@ -82,10 +82,10 @@ public class StockOrderDetailCustom {
 	public void setProduct_variant_current_inventory(BigInteger product_variant_current_inventory) {
 		this.product_variant_current_inventory = product_variant_current_inventory;
 	}
-	public Integer getProduct_variant_destination_inventory() {
+	public BigInteger getProduct_variant_destination_inventory() {
 		return product_variant_destination_inventory;
 	}
-	public void setProduct_variant_destination_inventory(Integer product_variant_destination_inventory) {
+	public void setProduct_variant_destination_inventory(BigInteger product_variant_destination_inventory) {
 		this.product_variant_destination_inventory = product_variant_destination_inventory;
 	}
 	public boolean isIS_PRODUCT() {
